+++ CHANGES.rst
@@ -6,7 +6,9 @@ Changelog for Websauna
 -------------------
 
 - Closes `Issue #179`_: Support for Celery 4.2.0.
+
 - Adds support to Python 3.7.0.
+
 - Reorganize tests.